Refactor completion request module
