+++ src-tauri/src/agent/llm/completion/mod.rs
@@ -16,10 +16,10 @@ pub use context::{
 };
 pub use orchestration::request_llm_completion_with_recovery;
 pub use request::{
-    build_compact_context_selection_options, build_compact_summary_message_for_messages,
-    build_compact_summary_text, merge_consecutive_user_messages, normalize_request_messages,
-    request_llm_completion, resolve_preserved_calibration_ratio,
-    try_apply_lossy_main_request_fallback,
+    build_compact_context_selection_options, build_compact_summary_message,
+    build_compact_summary_message_for_messages, build_compact_summary_text,
+    merge_consecutive_user_messages, normalize_request_messages, request_llm_completion,
+    resolve_preserved_calibration_ratio, try_apply_lossy_main_request_fallback,
 };
 
 // Crate-internal re-exports for intra-module visibility
